+++ config/test.exs
@@ -1,1 +1,3 @@
 import Config
+
+config :mockery, enable: true